+++ src/CleanRegex/Replace/Details/ReplaceDetail.php
@@ -82,9 +82,9 @@ class ReplaceDetail implements Detail
      * @param string|int $nameOrIndex
      * @return bool
      */
-    public function hasGroup($nameOrIndex): bool
+    public function groupExists($nameOrIndex): bool
     {
-        return $this->detail->hasGroup($nameOrIndex);
+        return $this->detail->groupExists($nameOrIndex);
     }
 
     public function text(): string