+++ test/Fakes/CleanRegex/Match/Details/ByteOffsetDetail.php
@@ -140,7 +140,7 @@ class ByteOffsetDetail implements Detail
         throw $this->fail();
     }
 
-    public function hasGroup($nameOrIndex): bool
+    public function groupExists($nameOrIndex): bool
     {
         throw $this->fail();
     }