+++ test/Fakes/CleanRegex/Match/Details/TextDetail.php
@@ -40,7 +40,7 @@ class TextDetail implements Detail
         throw $this->fail();
     }
 
-    public function hasGroup($nameOrIndex): bool
+    public function groupExists($nameOrIndex): bool
     {
         throw $this->fail();
     }